+++ .idea/runConfigurations/broker__dev_2_.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="broker (dev-2)" type="NodeJSConfigurationType" application-parameters="configs/development-2.env.json" path-to-js-file="bin/broker.js" working-dir="$PROJECT_DIR$/packages/broker">
+  <configuration default="false" name="broker (dev-2)" type="NodeJSConfigurationType" application-parameters="configs/development-2.env.json" path-to-js-file="dist/bin/broker.js" working-dir="$PROJECT_DIR$/packages/broker">
     <envs>
       <env name="LOG_LEVEL" value="info" />
     </envs>
