+++ .idea/runConfigurations/broker__dev_prod_resend_.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="broker (dev-prod-resend)" type="NodeJSConfigurationType" application-parameters="configs/development-prod-resend.env.json" path-to-js-file="bin/broker.js" working-dir="$PROJECT_DIR$/packages/broker">
+  <configuration default="false" name="broker (dev-prod-resend)" type="NodeJSConfigurationType" application-parameters="configs/development-prod-resend.env.json" path-to-js-file="dist/bin/broker.js" working-dir="$PROJECT_DIR$/packages/broker">
     <envs>
       <env name="LOG_LEVEL" value="info" />
     </envs>
