+++ packages/broker/.idea/runConfigurations/Run_Broker__8790__8791_.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Run Broker (8790, 8791)" type="NodeJSConfigurationType" application-parameters="configs/development-2.env.json" path-to-js-file="bin/broker.js" working-dir="$PROJECT_DIR$">
+  <configuration default="false" name="Run Broker (8790, 8791)" type="NodeJSConfigurationType" application-parameters="configs/development-2.env.json" path-to-js-file="dist/bin/broker.js" working-dir="$PROJECT_DIR$">
     <envs>
       <env name="LOG_LEVEL" value="debug" />
     </envs>
