+++ packages/network-tracker/bin/tracker.ts
@@ -1,22 +1,24 @@
 #!/usr/bin/env node
-/* eslint-disable @typescript-eslint/no-require-imports */
-const program = require('commander')
-const { startTracker } = require('../dist/src/startTracker')
-const { MetricsContext, Logger } = require('@streamr/utils')
-const ethers = require('ethers')
-const { SlackBot } = require('@streamr/slackbot')
-
-const CURRENT_VERSION = require('../package.json').version
+import { program } from 'commander'
+import pkg from '../package.json'
+import { startTracker } from '../src/startTracker'
+import { MetricsContext, Logger } from '@streamr/utils'
+import { Wallet } from 'ethers'
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore no declaration file for module
+import { SlackBot } from '@streamr/slackbot'
 
 const logger = new Logger(module)
 
+const parseIntOption = (value: string) => parseInt(value, 10)
+
 program
-    .version(CURRENT_VERSION)
+    .version(pkg.version)
     .usage('<ethereumPrivateKey>')
-    .option('--port <port>', 'port', 30300)
+    .option('--port <port>', 'port', parseIntOption, 30300)
     .option('--ip <ip>', 'ip', '0.0.0.0')
     .option('--unixSocket <unixSocket>', 'unixSocket', undefined)
-    .option('--maxNeighborsPerNode <maxNeighborsPerNode>', 'maxNeighborsPerNode', 4)
+    .option('--maxNeighborsPerNode <maxNeighborsPerNode>', 'maxNeighborsPerNode', parseIntOption, 4)
     .option('--attachHttpEndpoints', 'attach http endpoints')
     .option('--privateKeyFileName <privateKeyFileName>', 'private key filename', undefined)
     .option('--certFileName <certFileName>', 'cert filename', undefined)
@@ -25,28 +27,28 @@ program
     .option('--slackBotToken <slackBotToken>', 'slack API token', '')
     .option('--slackChannel <slackChannel>', 'slack channel for alerts', '#network-log')
 
-    .description('Run tracker with reporting')
+    .description('Run Streamr Tracker')
     .parse(process.argv)
 
 if (program.args.length < 1) {
     program.help()
 }
 const privateKey = program.args[0]
-const wallet = new ethers.Wallet(privateKey)
+const wallet = new Wallet(privateKey)
 const id = wallet.address
 const listen = program.opts().unixSocket ? program.opts().unixSocket : {
     hostname: program.opts().ip,
-    port: Number.parseInt(program.opts().port, 10)
+    port: program.opts().port
 }
 
 const { slackBotToken, slackChannel } = program.opts()
-let slackbot
+let slackbot: SlackBot
 const slackAlertHeader = `Tracker ${id}`
 if (slackBotToken && slackChannel) {
     slackbot = new SlackBot(slackChannel, slackBotToken)
 }
 
-const logError = (err, errorType) => {
+const logError = (err: any, errorType: string) => {
     logger.getFinalLogger().error(err, errorType)
     if (slackbot !== undefined) {
         const message = `${errorType}: ${err}`
@@ -72,7 +74,7 @@ async function main() {
         await startTracker({
             listen,
             id,
-            maxNeighborsPerNode: Number.parseInt(program.opts().maxNeighborsPerNode),
+            maxNeighborsPerNode: program.opts().maxNeighborsPerNode,
             attachHttpEndpoints: program.opts().attachHttpEndpoints,
             privateKeyFileName: program.opts().privateKeyFileName,
             certFileName: program.opts().certFileName,
@@ -81,7 +83,7 @@ async function main() {
             trackerPingInterval: 60 * 1000
         })
 
-        const trackerObj = {}
+        const trackerObj: any = {}
         const fields = [
             'ip', 'port', 'maxNeighborsPerNode', 'privateKeyFileName', 'certFileName', 'attachHttpEndpoints', 'unixSocket']
         fields.forEach((prop) => {
