+++ packages/network-tracker/package.json
@@ -10,7 +10,7 @@
   "license": "STREAMR NETWORK OPEN SOURCE LICENSE",
   "author": "Streamr Network AG <contact@streamr.network>",
   "bin": {
-    "streamr-tracker": "bin/tracker.js"
+    "streamr-tracker": "dist/bin/tracker.js"
   },
   "main": "dist/src/exports.js",
   "types": "dist/src/exports.d.ts",
